Move internal classes to package
