+++ core/src/main/java/org/apache/calcite/runtime/SpatialTypeFunctions.java
@@ -1878,62 +1878,4 @@ public class SpatialTypeFunctions {
       };
     }
   }
-
-  /**
-   * Used at run time by the ST_Union function.
-   */
-  public static class Union {
-
-    public List<Geometry> init() {
-      return new ArrayList<>();
-    }
-
-    public List<Geometry> add(List<Geometry> accumulator, Geometry geometry) {
-      accumulator.add(geometry);
-      return accumulator;
-    }
-
-    public Geometry result(List<Geometry> accumulator) {
-      return new UnaryUnionOp(accumulator).union();
-    }
-  }
-
-  /**
-   * Used at run time by the ST_Accum function.
-   */
-  public static class Accum {
-
-    public List<Geometry> init() {
-      return new ArrayList<>();
-    }
-
-    public List<Geometry> add(List<Geometry> accumulator, Geometry geometry) {
-      accumulator.add(geometry);
-      return accumulator;
-    }
-
-    public List<Geometry> result(List<Geometry> accumulator) {
-      return accumulator;
-    }
-  }
-
-  /**
-   * Used at run time by the ST_Collect function.
-   */
-  public static class Collect {
-
-    public List<Geometry> init() {
-      return new ArrayList<>();
-    }
-
-    public List<Geometry> add(List<Geometry> accumulator, Geometry geometry) {
-      accumulator.add(geometry);
-      return accumulator;
-    }
-
-    public Geometry result(List<Geometry> accumulator) {
-      Geometry[] array = accumulator.toArray(new Geometry[accumulator.size()]);
-      return GEOMETRY_FACTORY.createGeometryCollection(array);
-    }
-  }
 }
