+++ core/src/main/java/org/apache/calcite/runtime/UnionOperation.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to you under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.calcite.runtime;
+
+import org.locationtech.jts.geom.Geometry;
+import org.locationtech.jts.operation.union.UnaryUnionOp;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Used at run time by the ST_Union function.
+ */
+public class UnionOperation {
+
+  public List<Geometry> init() {
+    return new ArrayList<>();
+  }
+
+  public List<Geometry> add(List<Geometry> accumulator, Geometry geometry) {
+    accumulator.add(geometry);
+    return accumulator;
+  }
+
+  public Geometry result(List<Geometry> accumulator) {
+    return new UnaryUnionOp(accumulator).union();
+  }
+}
